requirements.txt
aiogram==2.19
requests==2.27.1
urllib3==1.26.9
python-dotenv==0.20.0